fastify fixes + docker rfks
