+++ hollowdb/config/.gitignore
@@ -1,0 +1,2 @@
+# hide yo wallet
+*.json
