+++ hollowdb_wait/wait.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Connection parameters
-TARGET=${TARGET:-"http://localhost:3000"}
+TARGET=${TARGET:-"http://localhost:3030"}
 
 # How long to sleep (seconds) before each attempt.
 SLEEPS=${SLEEPS:-2}
@@ -9,13 +9,11 @@ SLEEPS=${SLEEPS:-2}
 echo "Polling $TARGET"
 
 while true; do
-  # with --fail, it will fail for an error code >=400
-  # our micro API returns 503 when it is still refreshing
-  curl --fail --output /dev/null --silent --data '{"route": "STATE"}' "$TARGET"
+  curl --fail --silent "$TARGET/state"
 
   # check if exit code of curl is 0
   if [ $? -eq 0 ]; then
-    echo "HollowDB container is ready!"
+    echo "HollowDB API is ready!"
     exit 0
   fi
